+++ packages/roosterjs-editor-api/lib/table/insertTable.ts
@@ -1,6 +1,23 @@
 import { ChangeSource, IEditor, PositionType, TableFormat } from 'roosterjs-editor-types';
 import { Position, saveTableInfo, VTable } from 'roosterjs-editor-dom';
 
+const DEFAULT_FORMAT: Partial<TableFormat> = {
+    bgColor: null,
+    topBorderColor: '#ABABAB',
+    bottomBorderColor: '#ABABAB',
+    verticalBorderColor: '#ABABAB',
+    headerRow: false,
+    firstColumn: false,
+    bandedRows: false,
+    bandedColumns: false,
+    bgColumnColorEven: null,
+    bgColumnColorOdd: '#ABABAB20',
+    bgColorEven: null,
+    bgColorOdd: '#ABABAB20',
+    headerRowColor: '#ABABAB',
+    tableBorderFormat: null,
+};
+
 /**
  * Insert table into editor at current selection
  * @param editor The editor instance
@@ -35,17 +52,9 @@ export default function insertTable(
 
     editor.focus();
     editor.addUndoSnapshot(() => {
-        saveTableInfo(table, format);
         let vtable = new VTable(table);
-        vtable.applyFormat(
-            format || {
-                bgColorEven: '#FFF',
-                bgColorOdd: '#FFF',
-                topBorderColor: '#ABABAB',
-                bottomBorderColor: '#ABABAB',
-                verticalBorderColor: '#ABABAB',
-            }
-        );
+        saveTableInfo(table, format || DEFAULT_FORMAT);
+        vtable.applyFormat(format || DEFAULT_FORMAT);
         vtable.writeBack();
         editor.insertNode(fragment);
         editor.runAsync(editor =>
