+++ packages/roosterjs-editor-dom/lib/utils/tableInfo.ts
@@ -1,33 +1,16 @@
-import { format } from 'path';
+import toArray from './toArray';
 import { TableFormat } from 'roosterjs-editor-types';
-import { toArray } from '..';
 
 const TABLE_STYLE_INFO = 'roosterTableInfo';
-const DEFAULT_FORMAT: Partial<TableFormat> = {
-    bgColor: null,
-    topBorderColor: '#ABABAB',
-    bottomBorderColor: '#ABABAB',
-    verticalBorderColor: '#ABABAB',
-    headerRow: false,
-    firstColumn: false,
-    bandedRows: false,
-    bandedColumns: false,
-    bgColumnColorEven: null,
-    bgColumnColorOdd: '#ABABAB20',
-    bgColorEven: null,
-    bgColorOdd: '#ABABAB20',
-    headerRowColor: '#ABABAB',
-    tableBorderFormat: null,
-};
 
 /**
  * Save the format info of a table
  * @param table The table the info will be saved
  * @param format The format of the table
  */
-export function saveTableInfo(table: HTMLTableElement, format?: Partial<TableFormat>) {
-    if (table) {
-        table.dataset[TABLE_STYLE_INFO] = JSON.stringify(format ? format : DEFAULT_FORMAT);
+export function saveTableInfo(table: HTMLTableElement, format: Partial<TableFormat>) {
+    if (table && format) {
+        table.dataset[TABLE_STYLE_INFO] = JSON.stringify(format);
     }
 }
 
@@ -36,7 +19,7 @@ export function saveTableInfo(table: HTMLTableElement, format?: Partial<TableFor
  * @param table The table the info will be deleted
  */
 export function deleteTableInfo(table: HTMLTableElement) {
-    if (table && format) {
+    if (table) {
         delete table.dataset[TABLE_STYLE_INFO];
     }
 }
